BLOJSOM-250 Encode and Escape Technorati Tags produced by the TechnoratiTagsPlugin
The TechnoratiTagsPlugin class escapes the `tag' and link with a "simple" formatter provided by BlojsomUtils. The class should instead instead Encode and Escape the `tag'. Encode for the href and Escape for the "label" or "text" of the tag link. 

The change leverages BlojsomUtils in both cases, but modifies the String array provided to the message formatter to use the `tag' to produce both an encoded and escaped version of the tag. 

An additional modification was made to the technorati-tag-link Velocity template to use the new String array properly.