"Roughing out" the Maven 2 conversion for blojsom v3.5.
Generally, will build but not run as blojsom. Lots of work still to do with targets and the like. 

Questions about dependency management exist.

Questions about runtime library versions, availability of some older libraries used to build blojsom and updating to more current libraries.

Instructions on building from source, using Maven will need to be provided.

textile4j appears to be a dead project and the textile plugin will have to be updated with a new library.